+++ src/test/java/seedu/address/testutil/TestUtil.java
@@ -3,7 +3,6 @@ package seedu.address.testutil;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 import guitests.guihandles.PersonCardHandle;
@@ -13,16 +12,9 @@ import javafx.scene.Node;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
 import javafx.scene.input.KeyCombination;
-import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.commons.util.XmlUtil;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.model.tag.UniqueTagList;
 
 /**
  * A utility class for test cases.
@@ -33,40 +25,6 @@ public class TestUtil {
      * Folder used for temporary files created during testing. Ignored by Git.
      */
     public static final String SANDBOX_FOLDER = FileUtil.getPath("./src/test/data/sandbox/");
-
-    /**
-     * A Person array containing sample persons data.
-     */
-    public static final Person[] SAMPLE_PERSONS_ARRAY = makeSamplePersonsArray();
-
-    private static Person[] makeSamplePersonsArray() {
-        try {
-            //CHECKSTYLE.OFF: LineLength
-            return new Person[]{
-                new Person(new Name("Ali Muster"), new Phone("9482424"), new Email("hans@google.com"), new Address("4th street"), new UniqueTagList()),
-                new Person(new Name("Boris Mueller"), new Phone("87249245"), new Email("ruth@google.com"), new Address("81th street"), new UniqueTagList()),
-                new Person(new Name("Carl Kurz"), new Phone("95352563"), new Email("heinz@yahoo.com"), new Address("wall street"), new UniqueTagList()),
-                new Person(new Name("Daniel Meier"), new Phone("87652533"), new Email("cornelia@google.com"), new Address("10th street"), new UniqueTagList()),
-                new Person(new Name("Elle Meyer"), new Phone("9482224"), new Email("werner@gmail.com"), new Address("michegan ave"), new UniqueTagList()),
-                new Person(new Name("Fiona Kunz"), new Phone("9482427"), new Email("lydia@gmail.com"), new Address("little tokyo"), new UniqueTagList()),
-                new Person(new Name("George Best"), new Phone("9482442"), new Email("anna@google.com"), new Address("4th street"), new UniqueTagList()),
-                new Person(new Name("Hoon Meier"), new Phone("8482424"), new Email("stefan@mail.com"), new Address("little india"), new UniqueTagList()),
-                new Person(new Name("Ida Mueller"), new Phone("8482131"), new Email("hans@google.com"), new Address("chicago ave"), new UniqueTagList())
-            };
-            //CHECKSTYLE.ON: LineLength
-        } catch (IllegalValueException e) {
-            assert false;
-            // not possible
-            return null;
-        }
-    }
-
-    /**
-     * @return a list of sample persons data.
-     */
-    public static List<Person> getSamplePersonsDataAsList() {
-        return Arrays.asList(SAMPLE_PERSONS_ARRAY);
-    }
 
     /**
      * Appends the {@code fileName} to the sandbox folder path.
